refactor: move logic out of forms
